move service configuration into configuration package
