+++ kubernetes-discovery-client/src/main/java/io/micronaut/kubernetes/discovery/KubernetesServiceInstanceProvider.java
@@ -16,7 +16,7 @@
 package io.micronaut.kubernetes.discovery;
 
 import io.micronaut.discovery.ServiceInstance;
-import io.micronaut.kubernetes.client.v1.KubernetesServiceConfiguration;
+import io.micronaut.kubernetes.configuration.KubernetesServiceConfiguration;
 import org.reactivestreams.Publisher;
 
 import java.util.List;
